+++ src/collective/excelexport/datasources/eeafaceted/view.py
@@ -1,6 +1,7 @@
 from copy import copy
 import urllib
 
+
 class ExportUrl(object):
 
     def export_url(self):