+++ CHANGES.rst
@@ -2,7 +2,22 @@ Changelog
 =========
 
 
-0.1 (unreleased)
+1.2 (unreleased)
+----------------
+
+- We can define a dexterity adapter for just one field using field name as
+  adapter name.
+  [thomasdesvenain]
+
+
+1.1 (2014-06-19)
+----------------
+
+- Rename search policy excelexport.search to avoid conflict with 'search' view.
+  [thomasdesvenain]
+
+
+1.0 (2014-06-02)
 ----------------
 
 - Initial release.